# docs/src/apps/swagger.yaml
info:
  description: Nesis API
  version: 0.1.1
  title: Nesis API
paths:
...
  /v1/tasks:
    get:
      summary: Get all tasks available.
      parameters:
      - in: header
        name: Authorization
        schema:
          type: string
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tasks'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    post:
      summary: Creates a new task.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskReq'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRes'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /v1/tasks/{task_id}:
    get:
      summary: Get a single task by taskId.
      parameters:
      - in: header
        name: Authorization
        schema:
          type: string
        required: true
      - in: path
        name: taskId
        schema:
          type: string
        required: true
        description: The task id to get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRes'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    delete:
      summary: Delete a single task by taskId.
      parameters:
      - in: header
        name: Authorization
        schema:
          type: string
        required: true
      - in: path
        name: taskId
        schema:
          type: string
        required: true
        description: The task id to delete
      responses:
        '200':
          description: OK
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    put:
      summary: Creates a new task.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskReq'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRes'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
...
openapi: 3.0.2
components:
  schemas:
...
    TaskReq:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        type:
          type: string
        schedule:
          type: string
        parent_id:
          type: string
        definition:
          type: object
    TaskRes:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        type:
          type: string
        schedule:
          type: string
        parent_id:
          type: string
        definition:
          type: object
        id:
          type: string
        status:
          type: string
        create_date:
          type: string
          format: null
          pattern: null
        update_date:
          type: string
          format: null
          pattern: null
    Tasks:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaskRes'
        count:
          type: integer
...
    Message:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
...